Now building quad-fat; the 64-bit versions are somewhat experimental though
git-svn-id: http://svn.mulle-kybernetik.com/OCMock/trunk@26 12620d81-68e2-0310-ae40-e3f550779089

+++ Source/NSInvocationOCMAdditionsTests.m
@@ -318,7 +318,7 @@
 	// Give it an argument
 	NSData *data = [@"foo" dataUsingEncoding:NSUTF8StringEncoding];
 	const void *bytes = [[@"foo" dataUsingEncoding:NSUTF8StringEncoding] bytes];
-	unsigned int length = [data length];
+	NSUInteger length = [data length];
 	[invocation setArgument:&bytes atIndex:2];
 	[invocation setArgument:&length atIndex:3];
